api/requirements.txt
fastapi==0.92.0
nltk==3.7
pandas==1.4.4
python-decouple==3.7
requests==2.28.1
uvicorn==0.20.0
seaborn==0.12.1
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.8.1
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
